pyproject.toml
[project]
name = "turri-agent-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "db-dtypes>=1.4.3",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "fpdf2>=2.8.3",
    "google-adk>=1.3.0",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-bigquery-storage>=2.32.0",
    "greenlet>=3.2.3",
    "ipykernel>=6.29.5",
    "loguru>=0.7.3",
    "markdown>=3.8.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pgvector>=0.4.1",
    "pillow>=11.2.1",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pyyaml>=6.0.2",
    "redis>=6.2.0",
    "sqlmodel>=0.0.24",
    "tqdm>=4.67.1",
]
